+++ viteclientts/src/pages/VillaAnna/VillaAnnaHomePage.tsx
@@ -12,7 +12,6 @@ const HomePage = () => {
               flexDirection: 'column',
               alignItems: 'center',
               justifyContent: 'center',
-              p: 4,
             }}
           >
             <CustomButton